Create fizzBuzz
Print all the numbers from 1 to 100 with two exceptions. For numbers divisible by 3, print "Fizz" instead of the number, and for the numbers divisible by 5, print "Buzz" instead. 

When you have that working, modify your program to print "FizzBuzz" for numbers that are divisible by both 3 and 5. 

(Job Interview Question)